+++ SQLComponents/SQLQuery.h
@@ -94,24 +94,24 @@ public:
 
   // Set parameters for statement
   void SetParameter  (int p_num,SQLVariant*   p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (int p_num,long          p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameterUL(int p_num,unsigned long p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (int p_num,const char*         p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (int p_num,int           p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameterUL(int p_num,unsigned int  p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (int p_num,const char*   p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (int p_num,CString&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (int p_num,SQLDate&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (int p_num,SQLTime&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (int p_num,SQLTimestamp& p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (int p_num,const bcd&          p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (int p_num,const bcd&    p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
 
   void SetParameter  (SQLVariant*   p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (long          p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameterUL(unsigned long p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (const char*         p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (int           p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameterUL(unsigned int  p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (const char*   p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (CString&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (SQLDate&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (SQLTime&      p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
   void SetParameter  (SQLTimestamp& p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
-  void SetParameter  (const bcd&          p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
+  void SetParameter  (const bcd&    p_param,SQLParamType p_type = P_SQL_PARAM_INPUT);
 
   // SINGLE STATEMENT
