+++ UnitTest/Detail_cxh.cpp
@@ -39,18 +39,18 @@ static char THIS_FILE[] = __FILE__;
 
 // Bring the contents of the class to a SOAPMessage 
 BEGIN_XML_SERIALIZE(Detail)
-    CXO_XML_SERIALIZE(long,   m_id,         "id",         XDT_Integer);
-    CXO_XML_SERIALIZE(long,   m_mast_id,    "mast_id",    XDT_Integer);
-    CXO_XML_SERIALIZE(long,   m_line,       "line",       XDT_Integer);
+    CXO_XML_SERIALIZE(int,    m_id,         "id",         XDT_Integer);
+    CXO_XML_SERIALIZE(int,    m_mast_id,    "mast_id",    XDT_Integer);
+    CXO_XML_SERIALIZE(int,    m_line,       "line",       XDT_Integer);
     CXO_XML_SERIALIZE(CString,m_description,"description",XDT_String);
     CXO_XML_SERIALIZE(bcd,    m_amount,     "amount",     XDT_Decimal);
 END_XML_SERIALIZE
 
 // Read the contents of an object from a SOAPMessage
 BEGIN_XML_DESERIALIZE(Detail)
-  CXO_XML_DESERIALIZE(long,   m_id,         "id",         XDT_Integer);
-  CXO_XML_DESERIALIZE(long,   m_mast_id,    "mast_id",    XDT_Integer);
-  CXO_XML_DESERIALIZE(long,   m_line,       "line",       XDT_Integer);
+  CXO_XML_DESERIALIZE(int,    m_id,         "id",         XDT_Integer);
+  CXO_XML_DESERIALIZE(int,    m_mast_id,    "mast_id",    XDT_Integer);
+  CXO_XML_DESERIALIZE(int,    m_line,       "line",       XDT_Integer);
   CXO_XML_DESERIALIZE(CString,m_description,"description",XDT_String);
   CXO_XML_DESERIALIZE(bcd,    m_amount,     "amount",     XDT_Decimal);
 END_XML_DESERIALIZE
@@ -58,18 +58,18 @@ END_XML_DESERIALIZE
 //////////////////////////////////////////////////////////////////////////
 
 BEGIN_DBS_SERIALIZE(Detail)
-    CXO_DBS_SERIALIZE(long,   m_id,         "id",         XDT_Integer);
-    CXO_DBS_SERIALIZE(long,   m_mast_id,    "mast_id",    XDT_Integer);
-    CXO_DBS_SERIALIZE(long,   m_line,       "line",       XDT_Integer);
+    CXO_DBS_SERIALIZE(int,    m_id,         "id",         XDT_Integer);
+    CXO_DBS_SERIALIZE(int,    m_mast_id,    "mast_id",    XDT_Integer);
+    CXO_DBS_SERIALIZE(int,    m_line,       "line",       XDT_Integer);
     CXO_DBS_SERIALIZE(CString,m_description,"description",XDT_String);
     CXO_DBS_SERIALIZE(bcd,    m_amount,     "amount",     XDT_Decimal);
 END_DBS_SERIALIZE
 
 
 BEGIN_DBS_DESERIALIZE(Detail)
-  CXO_DBS_DESERIALIZE(long,   m_id,         "id",         XDT_Integer);
-  CXO_DBS_DESERIALIZE(long,   m_mast_id,    "mast_id",    XDT_Integer);
-  CXO_DBS_DESERIALIZE(long,   m_line,       "line",       XDT_Integer);
+  CXO_DBS_DESERIALIZE(int,    m_id,         "id",         XDT_Integer);
+  CXO_DBS_DESERIALIZE(int,    m_mast_id,    "mast_id",    XDT_Integer);
+  CXO_DBS_DESERIALIZE(int,    m_line,       "line",       XDT_Integer);
   CXO_DBS_DESERIALIZE(CString,m_description,"description",XDT_String);
   CXO_DBS_DESERIALIZE(bcd,    m_amount,     "amount",     XDT_Decimal);
 END_DBS_DESERIALIZE
